Provide the popup window HTML source
The new legacy mode depends on a popup window to facilitate
communication between App Inventor running in a browser and the
Companion running on a device.

At MIT this popup window's HTML is provided by the Rendezvous
server. The source code for the actual HTML is provided in the popup
directory in our sources (added here).

Change-Id: Iece99b5f73768fcc128a89b558013b416daaeb74