Inline BluetoothReflection functionality
BluetoothReflection was needed on versions of Android prior to Android
SDK 5. However, we now have a minimum SDK of 7, so the functionality
that was implemented by reflection can now be implemented directly
without issue. This also helps by letting the Android linting tools
verify that we have the right permissions for different functionality,
which is especially relevant given the changes in Android 12 to the
Bluetooth permission model.

Change-Id: I84b2f086c680b48a217da0e398a1847dafdd19da