Introducing The ChatBot Component
ChatBot works a lot like Translator. It uses a proxy run on the MIT
Campus. This version is designed around ChatGPT, which is called by
the proxy. Conversational continuity is managed by the proxy.

The component stores a UUID internally. When a new conversation
starts, this UUID is the empty string. The proxy will then return a
new UUID when it returns an answer from ChatGPT. By using this new
UUID in subsequent messages, the proxy provides the previous state as
well as the new question/input to ChatGPT.

The ResetConversation function sets this UUID back to the empty
string.

Change-Id: Ie27f3d4e2d6dd6303dc85d518793fc8f278a0432