Fix crashes reported by Google
Google reports out crash statistics in the Play Store console. If we
have too many crashes, Google threatens to negatively impact our store
listing. This change addresses the top two crashes we are seeing in
the console.

1. ConcurrentModificationException in the EventDispatcher.

   I was unable to reproduce this behavior. This fix synchronizes
   access to the HashMap, to guard against access from more then one
   thread. It also uses COW semantics for event dispatcher (@author
   Evan Patton) to guard against modification from the same thread.

2. Crash out of MediaUtil.

   I was able to reproduce this one by attempting to set an image to
   the string "\/\/" which results in an IllegalArgumentException out
   of URI.create(), at least on Android 10, which this is most
   reported.

   This change converts the IllegalArgumentException into an
   IOException which we handle (at least by logged it). More
   importantly, we don't crash the Companion on an IOException.

Future work might want to better handle exceptions on background
threads, where [2] happens. Perhaps by doing a Notification on the UI
thread. But that is a change for another time.

Change-Id: I954ef77b1d4ba63f43de04c2fa344b29dc36af23