Update Java Runtime to Java 17
This change updates appengine-web.xml to specify the Java 17
runtime. This is a requirement of Google if we wish to update our
system after October of 2024.

Note: You cannot yet compile App Inventor with Java 17, you still need
to use Java 11. You *may* need to use Java 17 with the "gcloud app
deploy" command and with the devserver. Non-ideal. A future change
will hopefully permit us to compile with Java 17. But for now the need
to update is getting urgent! (as we write this in September of 2024!).

Change-Id: Ie912d1b3651a36aee68fd1c802649ab6a12f6e42